padding and hover changes
tweaked the behavior of the work in the porfolio, there is padding between each work, the hover effect is now a box shadow instead of a border color change, captions now have pointer-events disabled by default so that the hyperlink in the portfolio images and the hover effects still work if the pointer is on the caption itself.

caption box sizing might need tweaking now, as they are still responsive but do not scale as intended. Might need to have the text size scale to achieve this, but I'm tired.